script/stream/requirement.txt
elasticsearch
pymongo
kafka-python
dnspython